+++ src/main/java/com/company/service/impl/UserServiceImpl.java
@@ -6,6 +6,8 @@ import com.company.repository.entity.User;
 import com.company.service.UserService;
 import com.company.service.dto.ObjectMapperService;
 import com.company.service.dto.UserDto;
+import com.company.service.exception.EntityNotFoundException;
+import com.company.service.exception.ValidateException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -40,22 +42,20 @@ public class UserServiceImpl implements UserService {
     public UserDto findById(Long id) {
         log.debug("Get user by id={} from database users", id);
         User user = userRepository.findById(id);
-        UserDto userDto = mapper.toDto(user);
-        if (userDto == null) {
-            throw new RuntimeException("User with id:" + id + " doesn't exist");
+        if (user == null) {
+            throw new EntityNotFoundException("User with id:" + id + " doesn't exist");
         }
-        return userDto;
+        return mapper.toDto(user);
     }
 
     @Override
     public UserDto getUserByEmail(String email) {
         log.debug("Get user by email={} from database users", email);
         User user = userRepository.getUserByEmail(email);
-        UserDto userDto = mapper.toDto(user);
-        if (userDto == null) {
-            throw new RuntimeException("User with email:" + email + " doesn't exist");
+        if (user == null) {
+            throw new EntityNotFoundException("User with email:" + email + " doesn't exist");
         }
-        return userDto;
+        return mapper.toDto(user);
     }
 
     @Override
@@ -70,7 +70,7 @@ public class UserServiceImpl implements UserService {
         log.debug("Get user by LastName={} from database users", lastName);
         List<User> users = userRepository.getUserByLastName(lastName);
         if (users == null) {
-            throw new RuntimeException("Users with lastname: " + lastName + " don't exist");
+            throw new EntityNotFoundException("Users with lastname: " + lastName + " don't exist");
         }
         return users.stream().map(mapper::toDto).toList();
     }
@@ -88,19 +88,28 @@ public class UserServiceImpl implements UserService {
     @Override
     public void delete(Long id) {
         log.debug("Delete user by id={} in database users", id);
+        if (!userRepository.delete(id)) {
+            throw new EntityNotFoundException("No entity with id: " + id);
+        }
         userRepository.delete(id);
     }
 
 
     private void validateCreate(UserDto user) {
         if (user == null) {
-            throw new RuntimeException("Books can't be empty...");
+            throw new ValidateException("Users can't be empty...");
+        }
+        if (user.getAge() <= 0) {
+            throw new ValidateException("Age is not valid. Age can't be less 0");
         }
     }
 
     private void validateUpdate(UserDto user) {
         if (user == null) {
-            throw new RuntimeException("Users can't be empty...");
+            throw new ValidateException("Users can't be empty...");
+        }
+        if (user.getAge() <= 0) {
+            throw new ValidateException("Age is not valid. Age can't be less 0");
         }
     }
 }
